manifests: drop the 'Load' prefix
The manifests packages load embedded manifests as implementation
detail.
The consumers of the package should not concern themselves on
how the package creates the objects.
To emphasize this, we drop the `Load` prefix from all the public
functions, and we make some (LoadObject) private.

A future extension will allow the package to load manifests from
the live filesystem, overriding the builtin ones.

Signed-off-by: Francesco Romani <fromani@redhat.com>