pkg: revamp logging
We need only two level of logging
- regular status messages
- debug messages (for troubleshooting)

So we build a minimal abstraction on top of the stdlib's `log`
package.

It's important to keep the distinction between logging and output.
The former is about status messages, which are useful to track
the program flow, and which are most used to track progress or
to understand what went wrong.
The latter is about the expected output from a program.

In our case:
- deploy/remove has logging, not output.
- render has output, may have logging (in the future)
- validate has logging and output

We should never mix logging and output on the same destination
(e.g on stdout)

Signed-off-by: Francesco Romani <fromani@redhat.com>