manifests: add functions to return waitable object
To generalize the components creation/deletion,
we add helper function to return not just the object
defintions, but a helper alongside them to be used,
if the client wishes so, to wait for the resource
just created to be ready.

This way the deploy/remove flow can (optionally) make sure to
sync and return only when everything is actually ready.

Signed-off-by: Francesco Romani <fromani@redhat.com>