sched: manifests: fix the namespace config
The scheduler plugin must look for the NodeResourceTopology
object in the right set of namespaces, otherwise it will
find no data.

Most notably, its configuration must include the namespace
on which the topology exporter creates objects into.

In this PR, we take this namespace from the manifests
of the topology exporter, and we make sure the scheduler
plugin configuration includes these values.

Signed-off-by: Francesco Romani <fromani@redhat.com>