Use NFD as the topology-updater component
So far we could only use RTE as the topology-updater component.
Now we can replace it with NFD.
NFD is the u/s version of RTE, thus it's is important to have the ability to use it.

The default topology-updater componenet is RTE since its more mature and tested for quite some time.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>