rte: omit namespace from core manifests
In order to simplify and generalize the manifest handling in
the RTE deployment flow, the core manifest set for RTE doesn't
include the namespace anymore.

In some important cases (deploying on kubernetes) a separate
namespace is still needed, and it is still provided, but must
be loaded and consumed explicitly.

Signed-off-by: Francesco Romani <fromani@redhat.com>