Fix DaemonSet update command
Consecutive calls for Manifests.Update() will append the same arg into the DaemonSet.Spec.Template.Spec.Containers[x].Command multiple time.
When we'll try to deploy the DaemonSet pods` the pod will failed because of this redundant args in the Command field.

In the usual case of running the deployer binary once and exit, this bug won't reproduce.
But on the numaresource-operator, we're calling the Manifests.Update() function multiple times and this is where the bug occoured.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>