Manifests.Update() will be called on fresh DaemonSet copy
The deployer's buisness logic was intended to get executed once per the deployer process's lifetime.
We reused the same logic on the numaresources-operator, but on the operator we're performing multiple updates on the same manifests instance - IOW the Update() function processed a dirty copy of DaemonSet object if more than a single Update() call being executed.

This lead to several bugs:
1. https://github.com/k8stopologyawareschedwg/deployer/pull/55
2. The export-namespace https://github.com/k8stopologyawareschedwg/deployer/blob/main/pkg/manifests/yaml/rte/daemonset.yaml#L20 command argument, not updated properly (the initial reason for PR)

In order to solve this issue we change the Update() logic to always work on a fresh copy of DaemonSet object, hence we won't carry former changes during the Update() call.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>