manifest: inline values in RTE ds
The RTE daemonset manifest was the only one which needed
variable expansion. This adds a dep to our code, complicates
the overall flow and is not really helpful, since updates
to the manifest is warranted anyway.

So we simplify the flow inlining the values, which we
can possibly change (or we never had provision to change
in the first place) anyway.

Signed-off-by: Francesco Romani <fromani@redhat.com>