manifests: introduce and use the flagcodec package
The deployer has the fairly rare need to manipulate
command lines; lacking well known package, we add
a simple and intentionally limited in-house package
to abstract this kind of manipulation, and we
use it in the manifest update flow.

Should we find suitable third-party packages to do the
task, we will promptly evaluate the switch.

Signed-off-by: Francesco Romani <fromani@redhat.com>