Mount configmap with the correct name
We're creating the ConfigMap object with a name that is based on the DaemonSet's name,
but when we're mounting the ConfigMap as a volume we're using a constant name that is not the real ConfigMap's name.

In addition, the ConfigMap initialization should be prior to the checking of its existence.

The result is that the pod is getting deployed without a mounted ConfigMap.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>