manifests: rte: split args from command
Per the documentation:
```
When you create a Pod, you can define a command and arguments for the containers that run in the Pod.
To define a command, include the `command` field in the configuration file. To define arguments for
the command, include the `args` field in the configuration file.
The command and arguments that you define cannot be changed after the Pod is created.
```
(see:
https://kubernetes.io/docs/tasks/inject-data-application/define-command-argument-container/
see also:
https://kubernetes.io/docs/tasks/inject-data-application/define-command-argument-container/#notes)

The most correct and precise form is thus to supply *both* `command` and
`args` - and that was the original intent of the manifest; conveying
everything into the `command` is a slightly abuse, even if this works.

Signed-off-by: Francesco Romani <fromani@redhat.com>