validate: overhaul: check cluster version
Overhaul the `validate` package to check and store the cluster version.
Breaks the API - will require a major bump.

In this change, we add support to check the cluster version, enabling
smarter validation check down the road (e.g. feature gates).
Now validation helpers can (and should) alter the validation state
accumulating intermediate results - while still returning their specific
validation results.

Same goes for the cluster version - it needs to be stored in the
Validator object, and possibly to be checked first. This is why
we introduce and recommend create functions.

Signed-off-by: Francesco Romani <fromani@redhat.com>