e2e: split deployment + validation test
Currently we have tests which do the overall deployment AND
perform the minimal validation on top of it.
Due to lack of CI resources, we always skip the validation, but
this also incorrectly marks the deployment as skipped.

Let's split the tests to get a better signal; while the
validation is expected to keep skipping, we can get a better
understanding from the deployment test.

Signed-off-by: Francesco Romani <fromani@redhat.com>