assets: notifier: use absolute path
Use absolute path for jq. Even though `/usr/bin/jq` is not always
guaranteed to be correct, it is still better than depending on $PATH
when running as OCI hook.

Plus, in the most common container OSes, the jq path _is_ actually
/usr.bin/jq. Point in case, this is especially true for OCP, the only
platform on which we deploy and recommend this hook anyway.

Signed-off-by: Francesco Romani <fromani@redhat.com>