pkg: commands: extract utilities
In order to facilitate integration of the `commands` package - which
was an original goal, extract helpers to register flags and parse
options. This is useful because subcommands all require the
CommonOptions, which are handled by the `root` command using
the code being factored out.

This change makes it easier to integrate the subcommands without
requiring anymore the root command.

Signed-off-by: Francesco Romani <fromani@redhat.com>