pkg: detect cluster version alongside platform
We will soon need to support version-specific configuration
(e.g. in MachineConfig).
To enable this, we add the concept of platform version (retroactively
renaming the current platform as "kind") all across the stack.

This is a API-breaking enablement PR only: consumers of this new
data will be added in future PRs.

X-API-Breaking: yes
Signed-off-by: Francesco Romani <fromani@redhat.com>