selinuxpolicy: updates to enable support for OCP 4.11+/K8s 1.24+
In container-selinux versions v2.186.0, v2.187.0 and v2.188.0, due to the updates here:
https://github.com/containers/container-selinux/commit/cf704e467c9b421f0c803f437b716e0b0315c1b6
selinuxpolicy updates are required (updating `container_runtime_t` to `kubelet_t`).

This is consumed in OCP 4.11 and K8s 1.24 and we need to make sure that the selinuxpolicy is
appropriately updated.

In addition to that, we need to continue supporting the older selinuxpolicy supplied in older
versions (e.g. 4.10) and that is the reason this patch is proposing an additional selinuxpolicy
and not completely replacing the original one.

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>