pkg: log: move from tlog to logr
We added `tlog` because we wanted a minimal abstraction over
basic logging facilities. This abstraction served us well,
but became a (albeit small) NIH factor which adds friction to
code reuse.

`logr` gives the features we need and more, is actually not
harder to use and unlocks a brave new world of interoperability.

In this PR we move from tlog to logr, sunsetting the former
for good. Unfortunately, turns out the actually more practical thing
is to make one giant move in one go.

Noteworthy changes:
- structured logs by default
- logs move from standard to debug logger

No change should have a lasting impact, and this is a breaking (=
version-bump requiring) change anyway.

Closes: #80
Closes: #110
Signed-off-by: Francesco Romani <fromani@redhat.com>