pkg: deployer: dissolve Helper
remove the Helper abstraction. Seemed a good idea, turned out
to lack focus and adding unnecessary abstraction.
We turn the good bits of Helper into a plain Data-only object
we use to pass around the context. Some methods are too shallow
and are just inlined; other are copied (and made private) in
the one and only calling site.

Closes: #113
Signed-off-by: Francesco Romani <fromani@redhat.com>