pkg: clientutil: commands: use env in `nodes`
Aiming to maximize the reuse of clients, we now require
a `deployer.Environment` in the `nodes` subpkg.
This decision is not straightforward, as we have now a kinda clumsy
dependency between `deployer` and `clientutil/nodes`. Since `nodes`
is a subpkg, the pros seems to outweight the cons, but this decision
can be revisited in the future.

Surely, one way or another, `clientutil/nodes` should reuse context,
logger and client (= the Env, in some shape or form).

Signed-off-by: Francesco Romani <fromani@redhat.com>