context: add and use explicit context in libraries
Up until now, the project libraries hardcoded `context.TODO()`
instead of requiring a context from the caller.
This is a bad practice and makes impossible to cancel the APIs, even
when the underlyingl libraries allowed to in turn accepting a context.

Fix the API to accept contexts, and pass them through as needed.

Signed-off-by: Francesco Romani <fromani@redhat.com>