wait: api: introduce the With() helper
The functions in the wait package accumulated a large set of arguments.
By introducing the Waiter object, created with the With() helper and
reorganizing the way we pass arguments, we can have nicer APIs with
actually less code passed around.

Example:
from:
```
_, err := wait.ForDeploymentComplete(ctx, cli, log, mf.DPController, wait.DefaultPollInterval, wait.DefaultPollTimeout)
```
to:
```
_, err := wait.With(cli, log).ForDeploymentComplete(ctx, mf.DPController)
```

Note Waiter is initialized with default values, wich can be changed
using the Timeout() and Interval() helper using a fluent API.

Example:
```
_, err := wait.With(cli, log).Timeout(1*time.Minute).Interval(10*time.Second).ForDeploymentComplete(ctx, mf.DPController)
```

Signed-off-by: Francesco Romani <fromani@redhat.com>