manifests: objectupdate: split updates.go in subpkgs
The manifests updating code gradually became a kitchen sink
holding all the update-related code with no real structure.
Split and reorganize the code in subpackages, mimicing and
extending the `pkg/manifests` hierarchy.

The new package `objectupdate` tree will sit logically on
top of the manifests tree - you need to have access to manifests
to be able to change them! so it's ok to depend on the `manifests`

This deceptively large PR is mostly code movement and trivial
renames. The end behavior is not expected to change.

Test coverage will be improved in future PRs.

Signed-off-by: Francesco Romani <fromani@redhat.com>