objectupdate introduce DaemonSetOptions
Instead of having per-layer struct, which leads to a lot
of copying and most notably unnecessary complexity, define
generic cross-layer options struct and use them pervasively.

Up until now the duplication proven to be cumbersome, for a change.
We will re-evaluate and inline later when/if there are clear benefits.

Along the way, use flagcodec to more robust argument processing
when updating DaemonSets.

Signed-off-by: Francesco Romani <fromani@redhat.com>