rte: consolidate manifests
the base template manifests should be mutated as little as possible.
So push down in the YAML definition as much as we can - and what
is not supposed to change anyway: RTE will ALWAYS need access to
sysfs and podresources socket.

Also, use the flagcodec package to
further reduce unnecessary changes.

Signed-off-by: Francesco Romani <fromani@redhat.com>