wait: add helpers for CRDs
Add wait functions for CRD created/deleted.
Creating and deleting CRDs is expected to be fast, but
not immediate. In CI envs, we can have flakes because
CRDs are not deleted fast enough and the attempt to
re-create thus fails, failing the test.

In all practical environments the operation should
be so fast that the extra gets/wait time should
be negligible - and in general we do this steps
very rarely, in the order of 1-2 times per *cluster* lifespan.

Considering the cost is expected to be negligible and
that the code is now more correct, we add and
use functions, which should also reduce e2e flakes.

Signed-off-by: Francesco Romani <fromani@redhat.com>