pkg: flagcodec: enable flag normalization
Following the conventions outlined in RATIONALE.md,
the pflag packages allows using both `--v` and `-v`
(double or single dashes for single-letter flags), but
the flagcodec package  wasn't behaving the same, treating
the two flags as different.

We add options to conform to our own rationale and to
be more compatible with pflag.

Signed-off-by: Francesco Romani <fromani@redhat.com>