pkg: consolidate options in a package
We intentionally added per-package Options with the intent
to decouple them and make fully independent, hoping to make the
packages as easy as possible to consume.

Unfortunately things quickly spiralled out of control, and nowadays
we have a relatively long chain of rote conversions to do to pass
parameters from the higher level packages down to the actual consumers.

This is starting to really hurt maintainability, so we move back the
pendulum centralizing the Options and reducing them (de-normalizing?).

Both approaches are debatable. Centralizing options has its own
drawbacks, but at the current state is the lesser evil.
This is the beginning of this work, and in future release we are likely
to review the approach again.

Signed-off-by: Francesco Romani <fromani@redhat.com>