sched: support for HA mode
wire in support for HA mode. Requires replicas to be set > 1,
mostly for backward compatibility and to give a way to
disable the feature for backward compatibility.

Adding e2e tests is challenging because of the limitations
of the free tier CI (annoying, but understandable).
Because of this it is deferred to a future PR.
The feature is turned off by default so existing tests
are a good gate to avoid regressions.

Signed-off-by: Francesco Romani <fromani@redhat.com>