images: overhaul the package, avoid implicit context
We had a bug on which the `images` command, little more than
a thin wrapper over the `images` package, didn't honor the
environment variables override.
The root cause is still unknown.

However, the package had a lot of unnecessary implicit baggage,
doing stuff at `init` time, using globals.

So we take the chance to do a massive cleanup and remove most
of the implicit. The API surface is now smaller and cleaner
and the flow is much more straightforward.

This is an API breakage but the users of this package were
already few and the updates are pretty simple.

Signed-off-by: Francesco Romani <fromani@redhat.com>